+++ tests/Router/RouteTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
 
 namespace Linna\Tests;
 
-use Linna\Http\Route;
+use Linna\Router\Route;
 use PHPUnit\Framework\TestCase;
 
 /**